+++ src/cli/csv_convert.rs
@@ -3,38 +3,6 @@ use std::fmt::Display;
 use std::path::Path;
 use std::str::FromStr;
 
-#[derive(Debug, Parser)]
-#[command(name="rcli",version,author,about,long_about=None)]
-pub struct Opts {
-    #[command(subcommand)]
-    pub cmd: Subcommand,
-}
-#[derive(Debug, Parser)]
-pub enum Subcommand {
-    #[command(name = "csv")]
-    Csv(CsvOpts),
-
-    #[command(name = "genpass", about = "Generate random password")]
-    GenPass(GenPassOpts),
-}
-#[derive(Debug, Parser)]
-pub struct GenPassOpts {
-    #[arg(short, long, default_value_t = 16)]
-    pub length: u8,
-
-    #[arg(long, default_value_t = true)]
-    pub uppercase: bool,
-
-    #[arg(long, default_value_t = true)]
-    pub lowercase: bool,
-
-    #[arg(long, default_value_t = true)]
-    pub number: bool,
-
-    #[arg(long, default_value_t = true)]
-    pub symbol: bool,
-}
-
 #[derive(Debug, Clone, Copy)]
 pub enum OutputFormat {
     Json,